requirements.txt
httpx>=0.27
playwright>=1.44
google-search-results>=2.4
dnspython>=2.6
beautifulsoup4>=4.12
lxml>=5.2
anthropic>=0.28
python-dotenv>=1.0
tqdm>=4.66
